client del

+++ client_del.sh
@@ -1,6 +1,11 @@
 #!/bin/bash
-echo -e "\nRemove procfile \n"
-echo "TRAVIS_BUILD_DIR"
-echo $TRAVIS_BUILD_DIR
-echo "$TRAVIS_BUILD_DIR"
-rm Procfile
+echo -e "\nReady to delete procfile \n"
+echo "TRAVIS_BUILD_DIR" + $TRAVIS_BUILD_DIR
+#CURR_DIR="/home/travis/build/jojozhuang/code-editor-react"
+CURR_DIR=$TRAVIS_BUILD_DIR
+CURR_DIR=`basename "$CURR_DIR"`
+echo "CURR_DIR = $CURR_DIR"
+if [ $CURR_DIR == "code-editor-react" ]; then
+  echo -e "\n Procfile is removed \n"
+  rm Procfile
+fi
